Add QueueReadiness, rename TestQueuesApiModule
