+++ src/domain/health/health.repository.ts
@@ -1,20 +1,23 @@
-import { Inject, Injectable } from '@nestjs/common';
 import { HealthEntity } from '@/domain/health/entities/health.entity';
 import { IHealthRepository } from '@/domain/health/health.repository.interface';
 import {
   CacheReadiness,
   ICacheReadiness,
 } from '@/domain/interfaces/cache-readiness.interface';
+import {
+  IQueueReadiness,
+  QueueReadiness,
+} from '@/domain/interfaces/queue-readiness.interface';
 import { ILoggingService, LoggingService } from '@/logging/logging.interface';
-import { IQueuesRepository } from '@/domain/queues/queues-repository.interface';
+import { Inject, Injectable } from '@nestjs/common';
 
 @Injectable()
 export class HealthRepository implements IHealthRepository {
   constructor(
-    @Inject(CacheReadiness) private readonly cacheService: ICacheReadiness,
     @Inject(LoggingService) private readonly loggingService: ILoggingService,
-    @Inject(IQueuesRepository)
-    private readonly queuesRepository: IQueuesRepository,
+    @Inject(CacheReadiness) private readonly cacheService: ICacheReadiness,
+    @Inject(QueueReadiness)
+    private readonly queuesApi: IQueueReadiness,
   ) {}
 
   async isReady(): Promise<HealthEntity> {
@@ -25,7 +28,7 @@ export class HealthRepository implements IHealthRepository {
       return HealthEntity.NOT_READY;
     }
 
-    return this.queuesRepository.isReady()
+    return this.queuesApi.isReady()
       ? HealthEntity.READY
       : HealthEntity.NOT_READY;
   }
