+++ src/domain/interfaces/queue-readiness.interface.ts
@@ -1,0 +1,8 @@
+export const QueueReadiness = Symbol('IQueueReadiness');
+
+export interface IQueueReadiness {
+  /**
+   * Checks if the configured queue consumer is connected.
+   */
+  isReady(): boolean;
+}
