+++ src/routes/email/email.controller.delete-email.spec.ts
@@ -29,7 +29,7 @@ import { TestEmailApiModule } from '@/datasources/email-api/__tests__/test.email
 import { IEmailApi } from '@/domain/interfaces/email-api.interface';
 import { accountBuilder } from '@/domain/account/entities/__tests__/account.builder';
 import { INestApplication } from '@nestjs/common';
-import { TestQueuesApiModule } from '@/datasources/queues/__tests__/test.queue-consumer.module';
+import { TestQueuesApiModule } from '@/datasources/queues/__tests__/test.queues-api.module';
 import { QueuesApiModule } from '@/datasources/queues/queues-api.module';
 
 describe('Email controller delete email tests', () => {
