+++ src/routes/email/email.controller.get-email.spec.ts
@@ -21,7 +21,7 @@ import { accountBuilder } from '@/domain/account/entities/__tests__/account.buil
 import { faker } from '@faker-js/faker';
 import { AccountDoesNotExistError } from '@/domain/account/errors/account-does-not-exist.error';
 import { getAddress } from 'viem';
-import { TestQueuesApiModule } from '@/datasources/queues/__tests__/test.queue-consumer.module';
+import { TestQueuesApiModule } from '@/datasources/queues/__tests__/test.queues-api.module';
 import { QueuesApiModule } from '@/datasources/queues/queues-api.module';
 
 describe('Email controller get email tests', () => {
