+++ src/routes/email/email.controller.resend-verification.spec.ts
@@ -18,7 +18,7 @@ import { accountBuilder } from '@/domain/account/entities/__tests__/account.buil
 import { verificationCodeBuilder } from '@/domain/account/entities/__tests__/verification-code.builder';
 import { faker } from '@faker-js/faker';
 import { getAddress } from 'viem';
-import { TestQueuesApiModule } from '@/datasources/queues/__tests__/test.queue-consumer.module';
+import { TestQueuesApiModule } from '@/datasources/queues/__tests__/test.queues-api.module';
 import { QueuesApiModule } from '@/datasources/queues/queues-api.module';
 
 const resendLockWindowMs = 100;
