+++ src/routes/health/health.controller.spec.ts
@@ -16,7 +16,7 @@ import { AccountDataSourceModule } from '@/datasources/account/account.datasourc
 import { TestAccountDataSourceModule } from '@/datasources/account/__tests__/test.account.datasource.module';
 import { IQueuesApiService } from '@/datasources/queues/queues-api.service.interface';
 import { QueuesApiModule } from '@/datasources/queues/queues-api.module';
-import { TestQueuesApiModule } from '@/datasources/queues/__tests__/test.queue-consumer.module';
+import { TestQueuesApiModule } from '@/datasources/queues/__tests__/test.queues-api.module';
 
 describe('Health Controller tests', () => {
   let app: INestApplication;
