+++ src/routes/transactions/__tests__/controllers/preview-transaction.transactions.controller.spec.ts
@@ -28,7 +28,7 @@ import { previewTransactionDtoBuilder } from '@/routes/transactions/entities/__t
 import { CacheModule } from '@/datasources/cache/cache.module';
 import { NetworkModule } from '@/datasources/network/network.module';
 import { getAddress } from 'viem';
-import { TestQueuesApiModule } from '@/datasources/queues/__tests__/test.queue-consumer.module';
+import { TestQueuesApiModule } from '@/datasources/queues/__tests__/test.queues-api.module';
 import { QueuesApiModule } from '@/datasources/queues/queues-api.module';
 
 describe('Preview transaction - Transactions Controller (Unit)', () => {
